refactor(ui): move tape_scan_blocks + route topbar_height — P1.5.2 step 2
Two related but small decouplings preparing the headless build target.

**tape_scan_blocks() → src/tape.cpp / src/tape.h.**  The function is a pure
TZX/CDT block walker — no ImGui calls, just reads the tape image and
populates `imgui_state.tape_block_offsets`.  Living in `imgui_ui.cpp`
meant the headless build (which excludes `imgui_ui.cpp` via the
`KONCPC_BUILD_MODERN_UI=OFF` gate from P1.5.1 step 4) couldn't call it.
Moving it next to the other `Tape_*` functions is also a natural home.

**`imgui_topbar_height()` routed through `IUiHost`.**  Added
`virtual int topbar_height() const = 0;` to the host interface.
NullUiHost returns 0 (headless mode renders the CPC viewport full-width).
ImGuiUiHost wraps the existing `imgui_topbar_height()`.  The two
callsites in `kon_cpc_ja.cpp`'s main loop now call `ui_host().topbar_height()`
instead of the bare free function, so they compile in MODERN_UI=OFF builds.

Two new unit tests:
- `IUiHostTest.NullHostTopbarHeightIsZero` — explicit `UiHostOverride(nullptr)`
  scope to test the null host's contract independently of whatever
  static-init installed as the process default.
- `IUiHostTest.OverrideReportsTopbarHeight` — RecordingHost returns its
  configured value.

Existing `NullHostIsDefault` test gains a comment explaining why it
doesn't assert on `topbar_height()` directly (modern-UI test builds
install ImGuiUiHost via AutoInstaller, which returns the ≥19px ImGui
menubar measurement).

Removes 2 of the 23 symbols that block MODERN_UI=OFF executable linking;
the remaining symbols (8 `video_*` helpers, `g_command_palette`,
`g_devtools_ui` + their methods, `imgui_mru_push`, `serial_terminal_feed_byte`,
`mainSDLWindow`, `video_plugin_list`, `g_repaint_*`) need bigger
decoupling that's better addressed by sub-PR 3 introducing the actual
`koncpc-core` executable with a slimmer main.

Modern-UI build is bit-identical; 1116/1118 tests pass (2 new + same 2
pre-existing RamExpansion skips).

bd: beads-6oa
