+++ src/imgui_ui.cpp
@@ -637,105 +637,8 @@ void imgui_close_menu()
   }
 }
 
-// ─────────────────────────────────────────────────
-// Tape block scanner — builds offset table from TZX image
-// ─────────────────────────────────────────────────
-
-// safe_read_word/dword moved to imgui_ui_testable.h
-
-void tape_scan_blocks()
-{
-  imgui_state.tape_block_offsets.clear();
-  imgui_state.tape_current_block = 0;
-  if (pbTapeImage.empty()) return;
-
-  byte* p = &pbTapeImage[0];
-  byte* end = pbTapeImageEnd;
-
-  while (p < end) {
-    imgui_state.tape_block_offsets.push_back(p);
-
-    // Calculate block size with bounds checking
-    // Same size logic as Tape_BlockDone + Tape_GetNextBlock
-    size_t block_size = 0;
-    word w; dword d;
-
-    switch (*p) {
-      case 0x10: // Standard speed data
-        if (!safe_read_word(p, end, 0x03, w)) goto done;
-        block_size = w + 0x04 + 1;
-        break;
-      case 0x11: // Turbo speed data
-        if (!safe_read_dword(p, end, 0x10, d)) goto done;
-        block_size = (d & 0x00ffffff) + 0x12 + 1;
-        break;
-      case 0x12: // Pure tone
-        block_size = 4 + 1;
-        break;
-      case 0x13: // Pulse sequence
-        if (p + 2 > end) goto done;
-        block_size = *(p+0x01) * 2 + 1 + 1;
-        break;
-      case 0x14: // Pure data
-        if (!safe_read_dword(p, end, 0x08, d)) goto done;
-        block_size = (d & 0x00ffffff) + 0x0a + 1;
-        break;
-      case 0x15: // Direct recording
-        if (!safe_read_dword(p, end, 0x06, d)) goto done;
-        block_size = (d & 0x00ffffff) + 0x08 + 1;
-        break;
-      case 0x20: // Pause
-        block_size = 2 + 1;
-        break;
-      case 0x21: // Group start
-        if (p + 2 > end) goto done;
-        block_size = *(p+0x01) + 1 + 1;
-        break;
-      case 0x22: // Group end
-        block_size = 1;
-        break;
-      case 0x30: // Text description
-        if (p + 2 > end) goto done;
-        block_size = *(p+0x01) + 1 + 1;
-        break;
-      case 0x31: // Message
-        if (p + 3 > end) goto done;
-        block_size = *(p+0x02) + 2 + 1;
-        break;
-      case 0x32: // Archive info
-        if (!safe_read_word(p, end, 0x01, w)) goto done;
-        block_size = w + 2 + 1;
-        break;
-      case 0x33: // Hardware type
-        if (p + 2 > end) goto done;
-        block_size = (*(p+0x01) * 3) + 1 + 1;
-        break;
-      case 0x34: // Emulation info
-        block_size = 8 + 1;
-        break;
-      case 0x35: // Custom info
-        if (!safe_read_dword(p, end, 0x11, d)) goto done;
-        block_size = d + 0x14 + 1;
-        break;
-      case 0x40: // Snapshot
-        if (!safe_read_dword(p, end, 0x02, d)) goto done;
-        block_size = (d & 0x00ffffff) + 0x04 + 1;
-        break;
-      case 0x5A: // Glue
-        block_size = 9 + 1;
-        break;
-      default: // Unknown block with 4-byte length
-        if (!safe_read_dword(p, end, 0x01, d)) goto done;
-        block_size = d + 4 + 1;
-        break;
-    }
-
-    // Validate we won't advance past end
-    if (p + block_size > end) goto done;
-    p += block_size;
-  }
-done:;
-}
+// tape_scan_blocks() moved to src/tape.cpp (declaration in tape.h) so the
+// headless build can call it without linking the modern UI.
 
 // ─────────────────────────────────────────────────
 // Main Menu Bar
